+++ src/core/BUILD.gn
@@ -387,6 +387,8 @@ openthread_core_files = [
   "border_router/dhcp6_pd_client.hpp",
   "border_router/infra_if.cpp",
   "border_router/infra_if.hpp",
+  "border_router/multi_ail_detector.cpp",
+  "border_router/multi_ail_detector.hpp",
   "border_router/routing_manager.cpp",
   "border_router/routing_manager.hpp",
   "border_router/rx_ra_tracker.cpp",
